Add files via upload
What I have so far - very procedural format to it, as I haven't quite wrapped my head around OO yet and this is my first brush with c#. Apologies for any issues that arise from this as a result. So far, the program takes in a pathname from the user for the spreadsheets , takes in the data from the spreadsheet, places it into a 2D list structure, and then scans it to see if the warning temperature (stored in a sparate file) has been exceeded. If so, the user will be notified. The user then gets an option to carry out a number of options, which I'll be continuing work on currently.